Crate penumbra_sdk_dex

Re-exports

pub use lp::action::PositionClose;

pub use lp::action::PositionOpen;

pub use lp::action::PositionWithdraw;

pub use swap::Swap;

pub use swap_claim::SwapClaim;

Modules

event

genesis

lp

state_key

swap

swap_claim

Structs

BatchSwapOutputData

CandlestickData

DexParameters

The configuration parameters for the DEX component.

DirectedTradingPair

DirectedUnitPair

A directed tuple of Units, similar to a DirectedTradingPair but embedding useful denom data.

SwapExecution

Contains the summary data of a trade, for client consumption.

TradingPair

The canonical representation of a tuple of asset [Id]s.

TradingPairVar

Represents a trading pair in R1CS.