penumbra_sdk_dex
Module lp

Re-exports

pub use metadata::PositionMetadata;

Modules

action

metadata

plan

position

view

Structs

BareTradingFunction

The data describing a trading function.

BuyOrder

Helper structure for constructing a Position expressing the desire to buy the desired value in exchange for the offered value.

LpNft

The denomination of an LPNFT tracking both ownership and state of a position.

Reserves

The reserves of a position.

SellOrder

Helper structure for constructing a Position expressing the desire to sell the offered value in exchange for the desired value.

TradingFunction